requirements.txt
discord.py==1.5.1
pip==21.0.1
PyMySQL==1.0.2
requests~=2.25.1